rfactory: move alogica de chaching do metodo find by id de Building para IdCacheableModel files: - app/Models/Building.php - app/Models/IdCacheableModel.php
